+++ src/main/java/com/github/project3/dto/chat/ChatUsersRequest.java
@@ -1,0 +1,10 @@
+package com.github.project3.dto.chat;
+
+import lombok.Data;
+
+@Data
+public class ChatUsersRequest {
+    private Integer id;
+    private String username;
+    private String password;
+}
